+++ ontospy/tests/test_methods.py
@@ -5,7 +5,7 @@ Unit test stub for ontosPy
 
 Run like this:
 
-:path/to/ontospyProject>python -m ontospy.tests.test_methods
+$ python -m ontospy.tests.test_methods
 
 """
 
@@ -20,6 +20,26 @@ from ..core.utils import *
 
 # sanity check
 print("-------------------\nOntoSpy ",  VERSION, "\n-------------------")
+
+
+
+class SampleCustomEntity(ontospy.RDF_Entity):
+
+    def __init__(self, uri, rdftype=None, namespaces=None, ext_model=False):
+        super(SampleCustomEntity, self).__init__(uri, rdftype, namespaces, ext_model)
+
+    def __repr__(self):
+        return "<SampleCustomEntity *%s*>" % ( self.uri)
+
+    def disjointWith(self):
+        """
+        Example: pull out disjoint with statements
+        """
+        pred = "http://www.w3.org/2002/07/owl#disjointWith"
+        return self.getValuesForProperty(pred)
+
+
+
 
 
 class TestMethods(unittest.TestCase):
@@ -47,6 +67,8 @@ class TestMethods(unittest.TestCase):
 				for el in c.instances:
 					print(el.uri, el.qname)
 					print(el.getValuesForProperty("http://www.w3.org/1999/02/22-rdf-syntax-ns#type"))
+		
+		printDebug("Test completed succesfully.\n", "green")
 
 
 	def test2(self):
@@ -61,13 +83,13 @@ class TestMethods(unittest.TestCase):
 			print("RDF:TYPE VALUES: ")
 			print(c.getValuesForProperty("http://www.w3.org/1999/02/22-rdf-syntax-ns#type"))
 
-
+		printDebug("Test completed succesfully.\n", "green")
 	
 	def test3(self):
 		"""
 		build_entity_from_uri
 		"""
-		printDebug("\n=================\nTEST 2: Checking the <build_entity_from_uri> method", "green")
+		printDebug("\n=================\nTEST 3: Checking the <build_entity_from_uri> method", "green")
 
 		e = self.o.build_entity_from_uri("http://www.co-ode.org/ontologies/pizza/pizza.owl#Germany")
 		print("URI: ", e)
@@ -75,8 +97,22 @@ class TestMethods(unittest.TestCase):
 		print("BEST LABEL: ", e.bestLabel())
 		print("RDF SOURCE: ")
 		print(e.rdf_source())
+		printDebug("Test completed succesfully.\n", "green")
 
 
+	def test4(self):
+		"""
+		build_entity_from_uri - SampleCustomEntity
+		"""
+		printDebug("\n=================\nTEST 4: Checking the <build_entity_from_uri> method using a SampleCustomEntity class ", "green")
+
+		e = self.o.build_entity_from_uri("http://www.co-ode.org/ontologies/pizza/pizza.owl#FruttiDiMare", SampleCustomEntity)
+		print("URI: ", e)
+		print("RDFTYPE: ", e.rdftype)
+		print("BEST LABEL: ", e.bestLabel())
+		print("OWL DISJOINT WITH: ")
+		print("\n".join([x for x in e.disjointWith()]))
+		printDebug("Test completed succesfully.\n", "green")
 	
 	
 	print("Success.\n")